move App into the components folder, general cleanup and tweaks
