+++ src/components/Articles/Article/Article.js
@@ -10,11 +10,11 @@ const Article = (props) => {
         <div className="Right">
           <div className="Author">By {props.article.author} on {props.article.publishedAt}</div>
           <div className="Title">
-            <a target="_blank" href={props.article.url}>{props.article.title}</a>
+            <a target="_blank" rel="noopener noreferrer" href={props.article.url}>{props.article.title}</a>
           </div>
           <div className="Description">{props.article.description}</div>
           <div className="ButtonContainer">
-            <a href={props.article.url} target="_blank" title={props.article.title} className="Button">Read This Article</a>
+            <a href={props.article.url} target="_blank" rel="noopener noreferrer" title={props.article.title} className="Button">Read This Article</a>
           </div>
         </div>
       </div>
